BAEL-1787 - add Builder directory. Move method cocde to that package. Add code for using builder with a class.
