+++ lombok/src/main/java/com/baeldung/lombok/builder/ClientBuilder.java
@@ -1,4 +1,4 @@
-package com.baeldung.lombok.intro;
+package com.baeldung.lombok.builder;
 
 import lombok.Builder;
