+++ lombok/src/main/java/com/baeldung/lombok/builder/Widget.java
@@ -1,0 +1,13 @@
+package com.baeldung.lombok.builder;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder(toBuilder = true)
+public class Widget {
+
+    private final String name;
+    private final int id;
+
+}
